requirements.txt
semantic-kernel==1.18.0
google-generativeai==0.8.3
python-dotenv==1.0.1
pydantic>=2.0,<2.10
httpx==0.28.1
fastapi==0.115.6
uvicorn==0.34.0